+++ CHANGELOG.md
@@ -1,4 +1,6 @@
-## 0.0.14
+## 0.0.18
+1. add thy-empty component;
+1. add thy store for state management.
 
 ## 0.0.13